code review: avoid exporting `getCommonProtocol` function
